# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RiceText 论坛与小说富文本 API
  version: 0.1.0
  description: 文档、章节、用户互动、附件权益和投票均使用共享契约，并由 SQLite 服务持久化实现。
servers:
  - url: http://localhost:8787
    description: 本地开发 API
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: ricetext_session
      description: OIDC 登录后由 Worker 设置的 HttpOnly session cookie。
...
paths:
  /api/documents/{documentId}:
    get:
      operationId: getDocument
      tags:
        - 文档
      summary: 读取当前文档
      description: |-
        读取最新不可变修订的 Tiptap JSON。任何论坛身份均可访问。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: documentId
          in: path
          required: true
          description: 文档稳定 ID，例如 demo-post
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
            description: 文档稳定 ID，例如 demo-post
          example: demo-post
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - {}
        - cookieSession: []
      responses:
        "200":
          description: 当前文档，例如 revision 为 1 的 demo-post。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  title:
                    type: string
                    minLength: 1
                    maxLength: 200
                  schemaVersion:
                    type: integer
                    exclusiveMinimum: 0
                    maximum: 9007199254740991
                  revision:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  savedAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                  content:
                    type: object
                    properties:
                      type:
                        type: string
                        const: doc
                      content:
                        default: []
                        maxItems: 20000
                        type: array
                        items:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - type
                      - content
                    additionalProperties: false
                required:
                  - id
                  - title
                  - schemaVersion
                  - revision
                  - savedAt
                  - content
                additionalProperties: false
        "404":
          description: 文档不存在。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 文档不存在。
    put:
      operationId: updateDocument
      tags:
        - 文档
      summary: 保存文档并创建修订
      description: |-
        需要 author 或 moderator。baseRevision 过期返回 409；相同 clientMutationId 重试返回首次结果且不重复建版。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: documentId
          in: path
          required: true
          description: 文档稳定 ID，例如 demo-post
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
            description: 文档稳定 ID，例如 demo-post
          example: demo-post
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - cookieSession: []
      requestBody:
        required: true
        description: 请求字段由共享 Zod 契约校验。
        content:
          application/json:
            schema:
              type: object
              properties:
                schemaVersion:
                  type: integer
                  exclusiveMinimum: 0
                  maximum: 9007199254740991
                baseRevision:
                  type: integer
                  minimum: 0
                  maximum: 9007199254740991
                clientMutationId:
                  type: string
                  minLength: 1
                  maxLength: 128
                  pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                content:
                  type: object
                  properties:
                    type:
                      type: string
                      const: doc
                    content:
                      default: []
                      maxItems: 20000
                      type: array
                      items:
                        type: object
                        description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                        additionalProperties: true
                  required:
                    - type
                    - content
                  additionalProperties: false
                chapterId:
                  type: string
                  minLength: 1
                  maxLength: 128
                  pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
              required:
                - schemaVersion
                - baseRevision
                - clientMutationId
                - content
              additionalProperties: false
      responses:
        "200":
          description: 幂等重试命中的既有修订。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  title:
                    type: string
                    minLength: 1
                    maxLength: 200
                  schemaVersion:
                    type: integer
                    exclusiveMinimum: 0
                    maximum: 9007199254740991
                  revision:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  savedAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                  content:
                    type: object
                    properties:
                      type:
                        type: string
                        const: doc
                      content:
                        default: []
                        maxItems: 20000
                        type: array
                        items:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - type
                      - content
                    additionalProperties: false
                required:
                  - id
                  - title
                  - schemaVersion
                  - revision
                  - savedAt
                  - content
                additionalProperties: false
        "201":
          description: 新建的不可变修订。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  title:
                    type: string
                    minLength: 1
                    maxLength: 200
                  schemaVersion:
                    type: integer
                    exclusiveMinimum: 0
                    maximum: 9007199254740991
                  revision:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  savedAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                  content:
                    type: object
                    properties:
                      type:
                        type: string
                        const: doc
                      content:
                        default: []
                        maxItems: 20000
                        type: array
                        items:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - type
                      - content
                    additionalProperties: false
                required:
                  - id
                  - title
                  - schemaVersion
                  - revision
                  - savedAt
                  - content
                additionalProperties: false
        "401":
          description: 未登录或会话已过期。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "403":
          description: 当前身份无编辑权限。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 当前身份无编辑权限。
        "409":
          description: 当前 revision 与 baseRevision 冲突，details.currentRevision 可用于刷新。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 当前 revision 与 baseRevision 冲突，details.currentRevision 可用于刷新。
        "422":
          description: 正文包含非法节点、属性或 URL。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 正文包含非法节点、属性或 URL。
...
  /api/documents/{documentId}/revisions:
    get:
      operationId: listRevisions
      tags:
        - 文档
      summary: 分页读取版本历史
      description: |-
        按 revision 倒序返回；cursor 使用上一页最后一项的 revision。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: documentId
          in: path
          required: true
          description: 文档稳定 ID，例如 demo-post
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
            description: 文档稳定 ID，例如 demo-post
          example: demo-post
        - name: cursor
          in: query
          required: false
          description: 查询参数 cursor
          schema:
            type: string
            pattern: ^[1-9]\d*$
        - name: limit
          in: query
          required: false
          description: 查询参数 limit
          schema:
            default: 20
            type: integer
            minimum: 1
            maximum: 100
          example: 20
        - name: chapterId
          in: query
          required: false
          description: 查询参数 chapterId
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - {}
        - cookieSession: []
      responses:
        "200":
          description: 不可变版本摘要页。
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        revision:
                          type: integer
                          minimum: 0
                          maximum: 9007199254740991
                        schemaVersion:
                          type: integer
                          exclusiveMinimum: 0
                          maximum: 9007199254740991
                        savedAt:
                          type: string
                          format: date-time
                          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                        authorId:
                          type: string
                          minLength: 1
                          maxLength: 128
                          pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                        authorName:
                          type: string
                        operation:
                          type: string
                          enum:
                            - seed
                            - update
                            - rollback
                            - suggestion
                            - steps
                        summary:
                          type: string
                        stepsSummary:
                          anyOf:
                            - type: string
                            - type: "null"
                        targetRevision:
                          anyOf:
                            - type: integer
                              minimum: 0
                              maximum: 9007199254740991
                            - type: "null"
                      required:
                        - revision
                        - schemaVersion
                        - savedAt
                        - authorId
                        - authorName
                        - operation
                        - summary
                        - stepsSummary
                        - targetRevision
                      additionalProperties: false
                  pageInfo:
                    type: object
                    properties:
                      nextCursor:
                        anyOf:
                          - type: string
                          - type: "null"
                    required:
                      - nextCursor
                    additionalProperties: false
                required:
                  - items
                  - pageInfo
                additionalProperties: false
        "404":
          description: 文档不存在。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 文档不存在。
  /api/documents/{documentId}/revisions/{revision}:
    get:
      operationId: getRevision
      tags:
        - 文档
      summary: 读取指定历史版本
      description: |-
        读取不可变 revision 的完整 Tiptap JSON，用于只读比较。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: documentId
          in: path
          required: true
          description: 文档稳定 ID，例如 demo-post
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
            description: 文档稳定 ID，例如 demo-post
          example: demo-post
        - name: revision
          in: path
          required: true
          description: 不可变 revision 编号
          schema:
            type: string
            pattern: ^\d+$
            description: 不可变 revision 编号
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - {}
        - cookieSession: []
      responses:
        "200":
          description: 指定历史 revision 的完整快照。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  title:
                    type: string
                    minLength: 1
                    maxLength: 200
                  schemaVersion:
                    type: integer
                    exclusiveMinimum: 0
                    maximum: 9007199254740991
                  revision:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  savedAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                  content:
                    type: object
                    properties:
                      type:
                        type: string
                        const: doc
                      content:
                        default: []
                        maxItems: 20000
                        type: array
                        items:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - type
                      - content
                    additionalProperties: false
                required:
                  - id
                  - title
                  - schemaVersion
                  - revision
                  - savedAt
                  - content
                additionalProperties: false
        "404":
          description: 文档或版本不存在。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 文档或版本不存在。
...